claude adds ability to init input.json
here's the prompt:
We need a new mode, called SecretInitialize. Add that to the mode enum
in args.rs. We additionally need a new item to add to struct Args,
let's call it secrets_init_filepath. That should be a readable file,
so use the check_readable_file logic for this param in the impl for
Args, if the mode passed was Mode::SecretInitialize. Then, in main.rs,
add a new match branch to deal with this mode. Then in mod secrets,
make a new file initialize.rs. In that file, follow basically the same
logic outlined in the shell script within secrets_retrive.md, except
write it in rust, no dropping to a shell. Within initialize.rs, we're
going to read all the files the user specified in
args.secrets_init_filepath, which is a json with format
[{filenm:value},{filenm:value},...]. For each file in that
input file, create a json entry like we specified in
secrets_retrieve.md. When you've processed them all, write the results
to a file named input.json, in the directory the user specified with
param secret_mode_input_json, which is also a required param for this
mode.